# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Laravel with library app",
        "description": "Implementation of Swagger with in Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/authors": {
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Create author",
                "description": "Create author",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\AuthorController::create",
                "requestBody": {
                    "description": "Author body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Author name",
                                        "type": "string",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Author array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "name": {
                                                "description": "Author name",
                                                "type": "string",
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_token": "token example"
                    }
                ]
            }
        },
        "/v1/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Return all books",
                "description": "Get all books",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\BookController::index",
                "parameters": [
                    {
                        "name": "author_id",
                        "in": "path",
                        "description": "Return books by author id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books arrray",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "name": {
                                                "description": "Book name",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "authors": {
                                                "description": "Array of book authors",
                                                "type": "array",
                                                "items": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create book",
                "description": "Create book",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\BookController::create",
                "requestBody": {
                    "description": "Book body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Book name",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "authors": {
                                        "description": "Array of book authors ids",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "items": {}
                                        },
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "name": {
                                                "description": "Book name",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "authors": {
                                                "description": "Array of book authors",
                                                "type": "array",
                                                "items": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_token": "token example"
                    }
                ]
            }
        },
        "/v1/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Return one book",
                "description": "Get one book by it's id",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\BookController::getById",
                "responses": {
                    "200": {
                        "description": "Book array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "name": {
                                                "description": "Book name",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "authors": {
                                                "description": "Array of book authors",
                                                "type": "array",
                                                "items": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update book",
                "description": "Update book",
                "operationId": "App\\Http\\Controllers\\Api\\V1\\BookController::update",
                "requestBody": {
                    "description": "Book body",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Book name",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "authors": {
                                        "description": "Array of book authors ids",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "items": {}
                                        },
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Id",
                                                "type": "integer",
                                                "nullable": false
                                            },
                                            "name": {
                                                "description": "Book name",
                                                "type": "string",
                                                "nullable": false
                                            },
                                            "authors": {
                                                "description": "Array of book authors",
                                                "type": "array",
                                                "items": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "nullable": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_token": "token example"
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}